+++ pkg/analyzer/lib/src/dart/scanner/reader.dart
@@ -1,8 +1,0 @@
-// Copyright (c) 2014, the Dart project authors. Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-export 'package:_fe_analyzer_shared/src/scanner/reader.dart'
-    show CharacterReader, CharSequenceReader, SubSequenceReader;
-
-/// This file is intentionally left blank except for exports.