+++ pkg/analyzer/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 10.1.0-dev
+
+* Internal changes only
+
 ## 10.0.2
 
 * Internal changes only